oh no

+++ src/main/java/Ozone/Pre/Download.java
@@ -56,7 +56,6 @@ public class Download implements Runnable {
         InputStream stream = null;
         FileOutputStream outputStream = null;
         File temp = new File(file.getParent(), System.currentTimeMillis() + ".temp");
-
         try {
             outputStream = new FileOutputStream(temp);
             // Open connection to URL.
@@ -124,6 +123,8 @@ public class Download implements Runnable {
             }
 
             try {
+                file.getParentFile().mkdirs();
+                if(temp.exists())
                 Files.copy(temp.toPath(), file.toPath());
             }catch (IOException e) {
                 e.printStackTrace();
